refactor: 사다리 생성 전략 클래스 및 인터페이스를 한 패키지로 묶음
