+++ src/main/java/model/strategy/BuildStrategy.java
@@ -1,4 +1,4 @@
-package model;
+package model.strategy;
 
 import java.util.List;
